Move `Page` type into new `page` submodule
